refactor globals.js and imports
