+++ e2e/pypi.spec.js
@@ -1,5 +1,6 @@
-import { test, Expect } from '../fixtures.js';
-import { packageReportTagName, tooltipSourceSelector } from '../src/globals.js';
+import { Expect, test } from '../fixtures.js';
+import { packageReportTagName } from '../src/globals.js';
+import { tooltipSourceSelector } from './selectors.js';
 
 test.describe('pypi', () => {
   test('Package Report is visible in PyPI.org', async ({ page }) => {
