+++ e2e/selectors.js
@@ -1,0 +1,1 @@
+export const tooltipSourceSelector = '.overlay-indicator__tooltip__sources > .overlay-indicator__tooltip__source';
